+++ src/Connection.php
@@ -8,7 +8,9 @@ use Psr\Log\LogLevel;
 use Throwable;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoConnection;
 use Yiisoft\Db\Driver\Pdo\PdoCommandInterface;
+use Yiisoft\Db\Mysql\Column\ColumnFactory;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
+use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
 use Yiisoft\Db\Transaction\TransactionInterface;
@@ -63,30 +65,23 @@ final class Connection extends AbstractPdoConnection
         return new Transaction($this);
     }
 
+    public function getColumnFactory(): ColumnFactoryInterface
+    {
+        return new ColumnFactory();
+    }
+
     public function getQueryBuilder(): QueryBuilderInterface
     {
-        return $this->queryBuilder ??= new QueryBuilder(
-            $this->getQuoter(),
-            $this->getSchema(),
-            $this->getServerInfo(),
-        );
+        return $this->queryBuilder ??= new QueryBuilder($this);
     }
 
     public function getQuoter(): QuoterInterface
     {
-        if ($this->quoter === null) {
-            $this->quoter = new Quoter('`', '`', $this->getTablePrefix());
-        }
-
-        return $this->quoter;
+        return $this->quoter ??= new Quoter('`', '`', $this->getTablePrefix());
     }
 
     public function getSchema(): SchemaInterface
     {
-        if ($this->schema === null) {
-            $this->schema = new Schema($this, $this->schemaCache);
-        }
-
-        return $this->schema;
+        return $this->schema ??= new Schema($this, $this->schemaCache);
     }
 }
