+++ tests/SchemaTest.php
@@ -14,7 +14,6 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Mysql\Column\ColumnBuilder;
-use Yiisoft\Db\Mysql\Column\ColumnFactory;
 use Yiisoft\Db\Mysql\Schema;
 use Yiisoft\Db\Mysql\Tests\Support\TestTrait;
 use Yiisoft\Db\Query\Query;
@@ -490,15 +489,6 @@ final class SchemaTest extends CommonSchemaTest
             'float_col' => '-12345.6789',
             'numeric_col' => '-33.22',
         ], $row);
-
-        $db->close();
-    }
-
-    public function testGetColumnFactory(): void
-    {
-        $db = $this->getConnection();
-
-        $this->assertInstanceOf(ColumnFactory::class, $db->getSchema()->getColumnFactory());
 
         $db->close();
     }
